pyproject.toml
[project]
name = "local-climate-response"
version = "0.1.0"
description = "A plugin for Geodesignhub that analyzes diagrams and designs to produce local-level heat and shade impact analytics"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "markupsafe>=3.0.2",
    "werkzeug>=3.1.3",
    "geojson>=3.1.0",
    "gunicorn>=23.0.0",
    "itsdangerous>=2.2.0",
    "requests>=2.32.3",
    "shapely>=2.0.7",
    "cligj>=0.7.2",
    "redis>=5.2.1",
    "rq>=2.1.0",
    "dacite>=1.9.2",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "geopandas>=1.0.1",
    "pybdshadow>=0.3.5",
    "numpy>=2.2.3",
    "pyproj>=3.7.1",
    "flask-babel>=4.0.0",
    "flask-wtf>=1.2.2",
    "bootstrap-flask>=2.4.1",
    "pyyaml>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.1.0",
    "psycopg2-binary>=2.9.10",
    "geoalchemy2>=0.17.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dashboard"]
